Add connections b/w components to show data from api response
Aslo refactored the code and fixed crashes incurred
Aslo upload default assets

+++ GoMovies/Models/MovieListModel.swift
@@ -20,6 +20,6 @@ struct MovieListModel: Decodable {
         let overview: String
         let posterPath: String
         let releaseDate: String
-        let voteAverage: Int
+        let voteAverage: Double
     }
 }
